feat: create get study api (#65)
