+++ src/apis/study/getStudyList.ts
@@ -5,7 +5,7 @@ import { END_POINTS } from "@/constants/api";
 import type { GetStudyListResponseType } from "@/types/study";
 
 export const getStudyList = async () => {
-  const { data } = await axiosInstance.get<GetStudyListResponseType>(END_POINTS.STUDY);
+  const { data } = await axiosInstance.get<GetStudyListResponseType>(END_POINTS.STUDY_LIST);
 
   return data;
 };
