+++ src/apis/study/postStudy.ts
@@ -9,5 +9,5 @@ export const postStudy = async (createStudyForm: CreateStudyFormRequestType) => 
     headers: { "Content-Type": "application/json" },
   };
 
-  return axiosInstance.post(END_POINTS.STUDY, createStudyForm, config);
+  return axiosInstance.post(END_POINTS.POST_STUDY, createStudyForm, config);
 };
