+++ src/components/Study/StudyList/StudyList.tsx
@@ -23,7 +23,7 @@ const StudyList = () => {
         <PageConter current={1} total={6} />
       </div>
       {studyListData.data.map((studyData) => (
-        <StudyCard studyData={studyData} />
+        <StudyCard studyData={studyData} key={studyData.studyId} />
       ))}
     </article>
   );
