+++ src/constants/api.ts
@@ -1,5 +1,7 @@
 export const BASE_URL = process.env.NEXT_PUBLIC_API_BASE_URL;
 
 export const END_POINTS = {
-  STUDY: "/api/study",
+  STUDY_LIST: "/api/study-list",
+  STUDY: (studyId: string) => `/api/study/${studyId}`,
+  POST_STUDY: `/api/study`,
 };
